Reorganize the MapObject code
