+++ js/src/index.js
@@ -1,15 +1,13 @@
 import './css/styles.css';
 
-// import all of geojs to get optional renderers
-// we could pick and choose what we want in the future
-import 'geojs';
-
 import Geonotebook from './Geonotebook';
 import MapObject from './MapObject';
 import * as jsonrpc from './jsonrpc';
+import * as map from './map';
 
 export {
   Geonotebook,
   MapObject,
-  jsonrpc
+  jsonrpc,
+  map
 };
