+++ js/src/map/index.js
@@ -1,0 +1,6 @@
+import ol from './ol';
+import geojs from './geojs';
+
+export {
+  ol, geojs
+};
